requirements.txt
Django==2.2.4
djangorestframework==3.10.2
django-filter==2.2.0